// src/components/file-upload/types.ts
export enum UploadTypes {
  REGISTRATION_CERTIFICATE = 'REGISTRATION_CERTIFICATE',
  BPRA_CERTIFICATE = 'BPRA_CERTIFICATE',
  MOU = 'MOU',
  DIRECTOR_ZANID = 'DIRECTOR_ZANID',
  TIN_CERTIFICATE = 'TIN_CERTIFICATE',
  ZRA_NUMBER = 'ZRA_NUMBER',
  INSPECTION_REPORT = 'INSPECTION_REPORT',
  INCINERATOR_PICTURE = 'INCINERATOR_PICTURE',
  BUILDING_MAP = 'BUILDING_MAP',
  TAEC_LICENSE = 'TAEC_LICENSE',

  // Hotel-specific types
  HOTEL_EXTERIOR = 'HOTEL_EXTERIOR',
  HOTEL_INTERIOR = 'HOTEL_INTERIOR',
  HOTEL_LOBBY = 'HOTEL_LOBBY',
  HOTEL_ROOM = 'HOTEL_ROOM',
  HOTEL_BATHROOM = 'HOTEL_BATHROOM',
  HOTEL_RESTAURANT = 'HOTEL_RESTAURANT',
  HOTEL_BAR = 'HOTEL_BAR',
  HOTEL_POOL = 'HOTEL_POOL',
  HOTEL_GYM = 'HOTEL_GYM',
  HOTEL_SPA = 'HOTEL_SPA',
  COMPLIANCE_DOCUMENTS = 'COMPLIANCE_DOCUMENTS',
  HOTEL_CONFERENCE = 'HOTEL_CONFERENCE',
  HOTEL_AMENITIES = 'HOTEL_AMENITIES',
  HOTEL_VIEW = 'HOTEL_VIEW',
  HOTEL_FLOOR_PLAN = 'HOTEL_FLOOR_PLAN',
  HOTEL_LOGO = 'HOTEL_LOGO',
  HOTEL_CERTIFICATE = 'HOTEL_CERTIFICATE',
  HOTEL_LICENSE = 'HOTEL_LICENSE',

  // Assessment Request Document Types
  BUILDING_PLAN_DOCUMENTS = 'BUILDING_PLAN_DOCUMENTS',
  OCCUPATIONAL_CERTIFICATE = 'OCCUPATIONAL_CERTIFICATE',
  OPERATING_LICENSES = 'OPERATING_LICENSES',
  EIA_REPORTS_AUDITS = 'EIA_REPORTS_AUDITS',
  DRAINAGE_SYSTEM_DOCS = 'DRAINAGE_SYSTEM_DOCS',
  ROOM_DESIGNATION_DOCS = 'ROOM_DESIGNATION_DOCS',
  SAFE_DEPOSIT_DOCS = 'SAFE_DEPOSIT_DOCS',
  HAND_WASH_BASIN_DOCS = 'HAND_WASH_BASIN_DOCS',
  WASH_ROOMS_DOCS = 'WASH_ROOMS_DOCS',
  WASTE_DISPOSAL_DOCS = 'WASTE_DISPOSAL_DOCS',
  SEWAGE_DISPOSAL_DOCS = 'SEWAGE_DISPOSAL_DOCS',
  SWIMMING_POOL_DOCS = 'SWIMMING_POOL_DOCS',
  SEWAGE_SYSTEM_DOCS = 'SEWAGE_SYSTEM_DOCS',
  WATER_SUPPLY_DOCS = 'WATER_SUPPLY_DOCS',
  SOLID_WASTE_DOCS = 'SOLID_WASTE_DOCS',
  VENTILATION_SYSTEM_DOCS = 'VENTILATION_SYSTEM_DOCS',
  KITCHEN_FACILITIES_DOCS = 'KITCHEN_FACILITIES_DOCS',
  FIRE_EXTINGUISHER_DOCS = 'FIRE_EXTINGUISHER_DOCS',
  ESCAPE_ROUTES_DOCS = 'ESCAPE_ROUTES_DOCS',
  SECURITY_MEASURES_DOCS = 'SECURITY_MEASURES_DOCS',
  LIGHTING_SYSTEM_DOCS = 'LIGHTING_SYSTEM_DOCS',
  EMERGENCY_EQUIPMENT_DOCS = 'EMERGENCY_EQUIPMENT_DOCS',
  ELECTRICITY_SUPPLY_DOCS = 'ELECTRICITY_SUPPLY_DOCS',
  TELECOMMUNICATION_DOCS = 'TELECOMMUNICATION_DOCS',
  LAUNDRY_SERVICES_DOCS = 'LAUNDRY_SERVICES_DOCS',
  INTERNET_SERVICES_DOCS = 'INTERNET_SERVICES_DOCS',
  TRANSPORT_SERVICES_DOCS = 'TRANSPORT_SERVICES_DOCS',
  QUALIFIED_STAFF_DOCS = 'QUALIFIED_STAFF_DOCS',
  HEALTH_SAFETY_DOCS = 'HEALTH_SAFETY_DOCS',
  VERMIN_PROOFING_DOCS = 'VERMIN_PROOFING_DOCS',
  FIRE_SAFETY_DOCS = 'FIRE_SAFETY_DOCS',
  SECURITY_SYSTEMS_DOCS = 'SECURITY_SYSTEMS_DOCS',
  COMMUNICATION_SYSTEMS_DOCS = 'COMMUNICATION_SYSTEMS_DOCS',
  ELECTRICAL_SAFETY_DOCS = 'ELECTRICAL_SAFETY_DOCS',
  FIRST_AID_DOCS = 'FIRST_AID_DOCS',
  QUALIFIED_MANAGEMENT_DOCS = 'QUALIFIED_MANAGEMENT_DOCS',
  QUALIFIED_MANAGEMENT_STAFF_DOCS = 'QUALIFIED_MANAGEMENT_STAFF_DOCS',
  QUALIFIED_HEADS_DOCS = 'QUALIFIED_HEADS_DOCS',
  QUALIFIED_DEPARTMENTAL_HEADS_DOCS = 'QUALIFIED_DEPARTMENTAL_HEADS_DOCS',
  HEALTH_MEDICAL_DOCS = 'HEALTH_MEDICAL_DOCS',
  HEALTH_MEDICAL_EXAM_DOCS = 'HEALTH_MEDICAL_EXAM_DOCS',
  EMERGENCY_EVACUATION_DOCS = 'EMERGENCY_EVACUATION_DOCS',
  INSURANCE_DOCS = 'INSURANCE_DOCS',

  // Assessor certificate types
  ASSESSOR_CERTIFICATE = 'ASSESSOR_CERTIFICATE',
  ASSESSOR_TRANSCRIPT = 'ASSESSOR_TRANSCRIPT',
  ASSESSOR_VERIFICATION_LETTER = 'ASSESSOR_VERIFICATION_LETTER',

  PROFILE_PHOTO = 'PROFILE_PHOTO',
}

export type ProgressMap = Map<string, number>;

export interface UploadedFile {
  id?: number;
  uuid?: string;
  name: string;
  fileSize: number;
  uploadType: string;
  fileType: string;
  size: string;
  base64?: string;
  facilityId?: number;
  hotelId?: number;
  roomTypeId?: number;
  index?: number;
  editing?: boolean;
  editName: string;
  mediaUrl?: string;
  description?: string;
  isDefault?: boolean;
}